pyproject.toml
[build-system]
requires = ["setuptools>=60", "setuptools-scm>=8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simpleDOAS"
version = "0.1.0"
description = "DOAS"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
'pandas',
'matplotlib==3.8.0',
'numpy==1.26.0',
'scipy==1.11.3',
'PyYAML==6.0.1',
    'netCDF4==1.5.7',
    'pyproj',
    'tqdm'
]

